Add files via upload
This project involves the development of a predictive model using logistic regression to determine the likelihood of survival for passengers on the Titanic. The project includes data preprocessing, feature engineering, and training of a logistic regression model on historical Titanic passenger data. Once trained, the model can predict whether a passenger would have survived based on input features such as age, gender, class, and boarding location.